moved extractor to monitoring module
