+++ src/ctapipe/monitoring/extractor.py
